# src2/aspnetcore/src/Org.OpenAPITools/Controllers/DefaultApi.cs
/*
 * Configuration Service API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.Http;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using Org.OpenAPITools.Attributes;
using Org.OpenAPITools.Models;

namespace Org.OpenAPITools.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class DefaultApiController : ControllerBase
    { 
        /// <summary>
        /// Get configurations
        /// </summary>
        /// <param name="pageSize"></param>
        /// <param name="pageToken">Paging token (opaque string, may be JSON-serialized internally)</param>
        /// <response code="200">OK</response>
        /// <response code="400">Validation error</response>
        [HttpGet]
        [Route("/api/configurations")]
        [ValidateModelState]
        [SwaggerOperation("GetConfigurations")]
        [SwaggerResponse(statusCode: 200, type: typeof(GetConfigurationsResponse), description: "OK")]
        public virtual IActionResult GetConfigurations([FromQuery (Name = "pageSize")][Required()][Range(1, 1000)]int pageSize, [FromQuery (Name = "pageToken")]string pageToken)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default);
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);
            string exampleJson = null;
            exampleJson = "{\n  \"entries\" : [ {\n    \"value\" : \"value\",\n    \"key\" : \"key\"\n  }, {\n    \"value\" : \"value\",\n    \"key\" : \"key\"\n  } ],\n  \"pageToken\" : \"pageToken\"\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<GetConfigurationsResponse>(exampleJson)
            : default;
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Set configurations
        /// </summary>
        /// <param name="setConfigurationsRequest"></param>
        /// <response code="200">OK</response>
        /// <response code="400">Validation error</response>
        [HttpPost]
        [Route("/api/configurations")]
        [Consumes("application/json")]
        [ValidateModelState]
        [SwaggerOperation("SetConfigurations")]
        public virtual IActionResult SetConfigurations([FromBody]SetConfigurationsRequest setConfigurationsRequest)
        {

            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);
            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            throw new NotImplementedException();
        }
    }
}