# src2/aspnetcore/src/Org.OpenAPITools/Models/SetConfigurationsRequest.cs
/*
 * Configuration Service API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 * Generated by: https://openapi-generator.tech
 */

using System;
using System.Linq;
using System.Text;
using System.Collections.Generic;
using System.ComponentModel;
using System.ComponentModel.DataAnnotations;
using System.Runtime.Serialization;
using Newtonsoft.Json;
using Org.OpenAPITools.Converters;

namespace Org.OpenAPITools.Models
{ 
    /// <summary>
    /// 
    /// </summary>
    [DataContract]
    public partial class SetConfigurationsRequest : IEquatable<SetConfigurationsRequest>
    {
        /// <summary>
        /// Gets or Sets Entries
        /// </summary>
        [Required]
        [DataMember(Name="entries", EmitDefaultValue=false)]
        public List<SetConfigurationEntry> Entries { get; set; }

        /// <summary>
        /// Returns the string presentation of the object
        /// </summary>
        /// <returns>String presentation of the object</returns>
        public override string ToString()
        {
            var sb = new StringBuilder();
            sb.Append("class SetConfigurationsRequest {\n");
            sb.Append("  Entries: ").Append(Entries).Append("\n");
            sb.Append("}\n");
            return sb.ToString();
        }

        /// <summary>
        /// Returns the JSON string presentation of the object
        /// </summary>
        /// <returns>JSON string presentation of the object</returns>
        public string ToJson()
        {
            return JsonConvert.SerializeObject(this, Formatting.Indented);
        }

        /// <summary>
        /// Returns true if objects are equal
        /// </summary>
        /// <param name="obj">Object to be compared</param>
        /// <returns>Boolean</returns>
        public override bool Equals(object obj)
        {
            if (obj is null) return false;
            if (ReferenceEquals(this, obj)) return true;
            return obj.GetType() == GetType() && Equals((SetConfigurationsRequest)obj);
        }

        /// <summary>
        /// Returns true if SetConfigurationsRequest instances are equal
        /// </summary>
        /// <param name="other">Instance of SetConfigurationsRequest to be compared</param>
        /// <returns>Boolean</returns>
        public bool Equals(SetConfigurationsRequest other)
        {
            if (other is null) return false;
            if (ReferenceEquals(this, other)) return true;

            return 
                (
                    Entries == other.Entries ||
                    Entries != null &&
                    other.Entries != null &&
                    Entries.SequenceEqual(other.Entries)
                );
        }

        /// <summary>
        /// Gets the hash code
        /// </summary>
        /// <returns>Hash code</returns>
        public override int GetHashCode()
        {
            unchecked // Overflow is fine, just wrap
            {
                var hashCode = 41;
                // Suitable nullity checks etc, of course :)
                    if (Entries != null)
                    hashCode = hashCode * 59 + Entries.GetHashCode();
                return hashCode;
            }
        }

        #region Operators
        #pragma warning disable 1591

        public static bool operator ==(SetConfigurationsRequest left, SetConfigurationsRequest right)
        {
            return Equals(left, right);
        }

        public static bool operator !=(SetConfigurationsRequest left, SetConfigurationsRequest right)
        {
            return !Equals(left, right);
        }

        #pragma warning restore 1591
        #endregion Operators
    }
}